+++ src/ccbr_actions/docs.py
@@ -18,7 +18,7 @@ from .versions import (
 from .actions import set_output
 
 
-def get_docs_version(release_args=""):
+def get_docs_version(release_args="", release_tag=None, strict_semver=True):
     """
     Get correct version and alias for documentation website.
 
@@ -27,6 +27,8 @@ def get_docs_version(release_args=""):
 
     Args:
         release_args (str, optional): Additional arguments to pass to the `gh release` GitHub CLI command (default is "").
+        release_tag (str, optional): The tag of the release (default: None). Set this when running this code from a github release event. Only set this if the tag is currently checked out.
+        strict_semver (bool): If True, the version string must match the full semantic versioning pattern. Otherwise, a relaxed format with only the major and minor components is allowed.
 
     Returns:
         tuple: A tuple containing:
@@ -43,16 +45,21 @@ def get_docs_version(release_args=""):
         >>> get_docs_version()
         ('1.0', 'latest')
     """
-    release_tag = get_latest_release_tag(args=release_args).lstrip("v")
-    if not release_tag:
+    latest_release_tag = get_latest_release_tag(args=release_args)
+    if not latest_release_tag:
         warnings.warn("No latest release found")
 
     release_hash = get_latest_release_hash(args=release_args)
     current_hash = get_current_hash()
 
+    if not release_tag:
+        release_tag = latest_release_tag
+
     if release_hash == current_hash:
         docs_alias = "latest"
-        docs_version = get_major_minor_version(release_tag)
+        docs_version = get_major_minor_version(
+            release_tag.lstrip("v"), strict_semver=strict_semver
+        )
     else:
         if not release_hash or is_ancestor(
             ancestor=release_hash, descendant=current_hash
@@ -63,16 +70,27 @@ def get_docs_version(release_args=""):
             raise ValueError(
                 f"The current commit hash {current_hash[:7]} is not a descendent of the latest release {release_tag} {release_hash[:7]}"
             )
+
+    if not docs_version:
+        raise ValueError(
+            f"The major & minor components of the version cannot be determined from the release tag: {release_tag}. Hint: does the release tag follow semantic versioning?"
+        )
+
     return docs_version, docs_alias
 
 
-def set_docs_version():
+def set_docs_version(release_args="", release_tag=None, strict_semver=True):
     """
     Set version and alias in GitHub environment variables for docs website action.
 
     This function retrieves the documentation version and alias using
     `get_docs_version` and sets them as environment variables in the GitHub
     Actions environment.
+
+    Args:
+        release_args (str, optional): Additional arguments to determine the release version. Defaults to an empty string.
+        release_tag (str, optional): Specific release tag to use for determining the version. Defaults to None.
+        strict_semver (bool, optional): Whether to enforce strict semantic versioning. Defaults to True.
 
     Raises:
         ValueError: If the current commit hash is not a descendant of the latest release.
@@ -84,7 +102,9 @@ def set_docs_version():
     Examples:
         >>> set_docs_version()
     """
-    version, alias = get_docs_version()
+    version, alias = get_docs_version(
+        release_args=release_args, release_tag=release_tag, strict_semver=strict_semver
+    )
     set_output("VERSION", version)
     set_output("ALIAS", alias)